fix: prevent underflow and add range validation for game mode numbers
- Fix unsigned underflow in AdjustGameModeNum() when decrementing from 0
- Add explicit range validation in SetGameModeNum()
- Add GameModeNum struct to encapsulate min/max values and bit‑masks for numeric fields
- Define NumberOfApples and NumberOfStones for readability
- Update game mode bitmask from `uint16_t` to `uint32_t` to provide more space for future flags and values

Fixes #8
Fixes #9

+++ ApplesGame/Game.h
@@ -24,7 +24,7 @@ namespace ApplesGame
 	{
 		// Global game data
 		GameState gameState{};
-		uint16_t gameMode = 0;
+		uint32_t gameMode = 0;
 		unsigned int score = 0;
 
 		//Resources
